chore: move integration service logic to users/v1
