+++ src/services/api/integrationsApi/index.ts
@@ -1,15 +1,11 @@
 import { AxiosResponse } from "axios";
 import Integration from "types/entities/Integration";
 import IntegrationImpact from "types/apiResponses/IntegrationImpact";
-import { apiGet, apiPost, apiPut } from "..";
+import { apiGet } from "..";
 
 const integrationsApi = {
   getIntegration: (id: number | string): Promise<AxiosResponse<Integration>> =>
     apiGet(`integrations/${id}`),
-
-  createIntegration: (data: any): Promise<AxiosResponse<Integration>> =>
-    apiPost("integrations", data),
-
   getIntegrationImpact: (
     id: number | string,
     startDate: string,
@@ -21,10 +17,6 @@ const integrationsApi = {
         end_date: endDate,
       },
     }),
-  updateIntegration: (
-    id: any,
-    data: Integration,
-  ): Promise<AxiosResponse<Integration>> => apiPut(`integrations/${id}`, data),
 };
 
 export default integrationsApi;
